# backend/api/gen/go/admin/service/v1/i_language_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_language.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-admin/api/gen/go/dict/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationLanguageServiceBatchCreate = "/admin.service.v1.LanguageService/BatchCreate"
const OperationLanguageServiceCreate = "/admin.service.v1.LanguageService/Create"
const OperationLanguageServiceDelete = "/admin.service.v1.LanguageService/Delete"
const OperationLanguageServiceGet = "/admin.service.v1.LanguageService/Get"
const OperationLanguageServiceList = "/admin.service.v1.LanguageService/List"
const OperationLanguageServiceUpdate = "/admin.service.v1.LanguageService/Update"

type LanguageServiceHTTPServer interface {
	// BatchCreate 批量创建语言
	BatchCreate(context.Context, *v11.BatchCreateLanguagesRequest) (*emptypb.Empty, error)
	// Create 创建语言
	Create(context.Context, *v11.CreateLanguageRequest) (*emptypb.Empty, error)
	// Delete 删除语言
	Delete(context.Context, *v11.DeleteLanguageRequest) (*emptypb.Empty, error)
	// Get 查询语言详情
	Get(context.Context, *v11.GetLanguageRequest) (*v11.Language, error)
	// List 分页查询语言列表
	List(context.Context, *v1.PagingRequest) (*v11.ListLanguageResponse, error)
	// Update 更新语言
	Update(context.Context, *v11.UpdateLanguageRequest) (*emptypb.Empty, error)
}

func RegisterLanguageServiceHTTPServer(s *http.Server, srv LanguageServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/dict/langs", _LanguageService_List5_HTTP_Handler(srv))
	r.GET("/admin/v1/dict/langs/{id}", _LanguageService_Get5_HTTP_Handler(srv))
	r.POST("/admin/v1/dict/langs", _LanguageService_Create3_HTTP_Handler(srv))
	r.PUT("/admin/v1/dict/langs/{id}", _LanguageService_Update3_HTTP_Handler(srv))
	r.DELETE("/admin/v1/dict/langs", _LanguageService_Delete3_HTTP_Handler(srv))
	r.POST("/admin/v1/dict/langs/batch", _LanguageService_BatchCreate0_HTTP_Handler(srv))
}

func _LanguageService_List5_HTTP_Handler(srv LanguageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLanguageServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListLanguageResponse)
		return ctx.Result(200, reply)
	}
}

func _LanguageService_Get5_HTTP_Handler(srv LanguageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetLanguageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLanguageServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetLanguageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Language)
		return ctx.Result(200, reply)
	}
}

func _LanguageService_Create3_HTTP_Handler(srv LanguageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateLanguageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLanguageServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateLanguageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _LanguageService_Update3_HTTP_Handler(srv LanguageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateLanguageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLanguageServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateLanguageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _LanguageService_Delete3_HTTP_Handler(srv LanguageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteLanguageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLanguageServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteLanguageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _LanguageService_BatchCreate0_HTTP_Handler(srv LanguageServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.BatchCreateLanguagesRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLanguageServiceBatchCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BatchCreate(ctx, req.(*v11.BatchCreateLanguagesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type LanguageServiceHTTPClient interface {
	// BatchCreate 批量创建语言
	BatchCreate(ctx context.Context, req *v11.BatchCreateLanguagesRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Create 创建语言
	Create(ctx context.Context, req *v11.CreateLanguageRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Delete 删除语言
	Delete(ctx context.Context, req *v11.DeleteLanguageRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 查询语言详情
	Get(ctx context.Context, req *v11.GetLanguageRequest, opts ...http.CallOption) (rsp *v11.Language, err error)
	// List 分页查询语言列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListLanguageResponse, err error)
	// Update 更新语言
	Update(ctx context.Context, req *v11.UpdateLanguageRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type LanguageServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewLanguageServiceHTTPClient(client *http.Client) LanguageServiceHTTPClient {
	return &LanguageServiceHTTPClientImpl{client}
}

// BatchCreate 批量创建语言
func (c *LanguageServiceHTTPClientImpl) BatchCreate(ctx context.Context, in *v11.BatchCreateLanguagesRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/dict/langs/batch"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLanguageServiceBatchCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Create 创建语言
func (c *LanguageServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateLanguageRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/dict/langs"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLanguageServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除语言
func (c *LanguageServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteLanguageRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/dict/langs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLanguageServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 查询语言详情
func (c *LanguageServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetLanguageRequest, opts ...http.CallOption) (*v11.Language, error) {
	var out v11.Language
	pattern := "/admin/v1/dict/langs/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLanguageServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 分页查询语言列表
func (c *LanguageServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListLanguageResponse, error) {
	var out v11.ListLanguageResponse
	pattern := "/admin/v1/dict/langs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLanguageServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新语言
func (c *LanguageServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateLanguageRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/dict/langs/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLanguageServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}