# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/dict/langs:
        get:
            tags:
                - LanguageService
            description: 分页查询语言列表
            operationId: LanguageService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListLanguageResponse'
        post:
            tags:
                - LanguageService
            description: 创建语言
            operationId: LanguageService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateLanguageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - LanguageService
            description: 删除语言
            operationId: LanguageService_Delete
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dict/langs/batch:
        post:
            tags:
                - LanguageService
            description: 批量创建语言
            operationId: LanguageService_BatchCreate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BatchCreateLanguagesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dict/langs/{id}:
        get:
            tags:
                - LanguageService
            description: 查询语言详情
            operationId: LanguageService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: code
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Language'
        put:
            tags:
                - LanguageService
            description: 更新语言
            operationId: LanguageService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateLanguageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        BatchCreateLanguagesRequest:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/Language'
...
        CreateLanguageRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Language'
            description: 创建语言 - 请求
...
        Language:
            type: object
            properties:
                id:
                    type: integer
                    description: 语言ID
                    format: uint32
                languageCode:
                    type: string
                    description: 标准语言代码
                languageName:
                    type: string
                    description: 语言名称
                nativeName:
                    type: string
                    description: 本地语言名称
                isDefault:
                    type: boolean
                    description: 是否为默认语言
                isEnabled:
                    type: boolean
                    description: 是否启用
                sortOrder:
                    type: integer
                    description: 排序顺序，值越小越靠前
                    format: uint32
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 语言
...
        ListLanguageResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Language'
                total:
                    type: string
            description: 语言列表 - 答复
...
        UpdateLanguageRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/Language'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新语言 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...